+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - `unescape` property to label widget.
 - Better Nbcl error reporting.
 - `register_library` API to plugin API.
+- `emit` plugin API to emit signal to a shared buffer.
+- `listen` plugin API to listen to signals emitted by plugins.
 
 ### Changed